Rename flashloader/flashloader.[c|h] to flashloader/main.[c|h]
