+++ src/flashloader/main.c
@@ -1,4 +1,4 @@
-#include "flashloader.h"
+#include "main.h"
 
 #include "crc32.c"
