+++ src/flashloader/main.cmake
@@ -11,7 +11,7 @@ target_compile_definitions(${MAIN_FL_LINKED} PRIVATE INCLUDES_FLASHLOADER)
 
 target_sources(${MAIN_FL_LINKED} PUBLIC
     ${CMAKE_CURRENT_LIST_DIR}/crc32.c
-    ${CMAKE_CURRENT_LIST_DIR}/flashloader.c
+    ${CMAKE_CURRENT_LIST_DIR}/main.c
     ${CMAKE_CURRENT_LIST_DIR}/dfu.c
     ${CMAKE_CURRENT_LIST_DIR}/ihex/record.c
     ${CMAKE_CURRENT_LIST_DIR}/ihex/util.c
